+++ Changelog.md
@@ -4,6 +4,7 @@
 #### 2022-XX-YY
 
 - Added Components based on the Laravel Blade Components Feature
+- Added `page-actions` blade stack that themes need to define
 - Renamed the stylesheet of the default theme from `appshell` to `appshell3`
 - Changed the AppShell Theme - it has been rewritten from scratch with new visual style
 - Changed Breadcrumbs dependency from v7 to v8